Initial Commit
The first challenge that I worked on. The task was to find the perimeter of an island in a 2d Vertex. I changed it up to make it "harder" by using a 2d array instead. I figured it would be quicker and slightly faster. The islands are represented by 1's and water is represented by 0's